change libspdm_start_watchdog uint64_t seconds to uint16_t.

+++ include/hal/library/platform_lib.h
@@ -22,7 +22,7 @@ void libspdm_sleep(uint64_t milliseconds);
  * @param  seconds        heartbeat period, in seconds.
  *
  **/
-bool libspdm_start_watchdog(uint32_t session_id, uint64_t seconds);
+bool libspdm_start_watchdog(uint32_t session_id, uint16_t seconds);
 
 /**
  * stop watchdog.
